Update gtest and gmock to the latest versions.
In b6a0183ccec7, gtest and gmock were held back at older versions with
limited C++11 support, because the then-current heads of each did not
work well in environments with C++11 language support but pre-C++11
library support. This is the environment that Chrome code for Mac builds
in. All necessary patches for this have been merged upstream, and this
has been corrected in the current released versions of gtest and gmock.

TEST=*
R=rsesek@chromium.org

Review URL: https://codereview.chromium.org/731473007